[ISSUE 67] - Infinite loop when notifying model delegates of load start when the model is outdated
Part 1:
  Set the default cache expiration date for TTURLRequests to TT_DEFAULT_CACHE_EXPIRATION_AGE.

Part 2 (and why it's necessary):
The infinite loop occurs specifically when the model is outdated and the cache expirationdate
is set to 0. This causes a problem in
- (NSData*)dataForKey:(NSString*)key expires:(NSTimeInterval)expirationAge
   timestamp:(NSDate**)timestamp
because it will only return a miss on the data if expirationAge is non-zero. This doesn't make
intuitive sense because if the expirationAge is 0, any file should technically result in
a miss and be re-requested. This change considers expirationAge properly.

Paired with the default expiration age and the fix to cache misses, the cache should now properly
execute a miss on expired data without being excessive (the default expiration age is 1 week).