[ISSUE 85] - TTLauncherView - Switching to editing mode when there's only one TTLauncherItem
Fix sleuthed by github.com/uprise78, amended to avoid dangling UIView animations.

The problem:
UIView animations will immediately call the delegate on commitAnimations if nothing was actually
animated.

This caused a problem in the launcherview due to the fact that -wobble used itself as the
delegate to repeat the animation, causing infinite recursion when the launcher view was empty.
This fix counts the number of affected views and then only adds the -wobble delegate to the
animation if any views will actually be animated.

This differs from uprise78's original fix in that we now always commit the animation to avoid
unintentional nested UIView animation calls.

The original fix can be found here:
http://github.com/facebook/three20/issues#issue/85