[UnitTests] Add unit testing target to Three20.
From this point on we will be building unit tests for Three20, starting
with the core library components. The unit test target will be marked
as a dependency of the Three20 library so that every time you build Three20,
the unit tests also run.

There isn't a strict concept of "running" a unit test in Xcode. Instead you
simply build the unit test target and the tests are performed automatically.
If any tests fail, you will receive a build error-like notification with a
descriptive message.

New Unit Tests
* NSDataAdditions
  -(NSString*)md5hash;