Add TTHighlightBorderStyle and improve TTReflectiveFillStyle's algorithm.
TTHighlightBorderStyle is a border with a semi-transparent highlight
border drawn one pixel down and beneath the outer border. This makes
it possible to simulate the style of the "Delete Contact" button found in
the Contacts app.

TTReflectiveFillStyle's algorithm was just plain broken; it used a mashup of
hand-picked HSV numbers to "tint" the edge of the gradient. This change simply
draws a semi-transparent white gradient on top of the fill color. It also
adds a subtle, semi-transparent gradient to the bottom of the fill to complete
the bubble illusion.

Added:
+ (TTReflectiveFillStyle*)styleWithColor:(UIColor*)color
                          withBottomHighlight:(BOOL)withBottomHighlight
                          next:(TTStyle*)next

If the original styleWithColor:next: method is used, withBottomHighlight
defaults to NO.

Testing: Open TTCatalog and check out the styled views. The two nav-bar styles
should still have a nice gradient.