[BugFix] Fix NSArray's containsObject method.
performSelector: will return a value of type id. This caused difficult
to repro bugs in containsObject's use of performSelector: with methods
that return type BOOL (an unsigned char). This was because the method
would return an unsigned char, but this would then be cast to id
(a four-byte variable). This left room for uninitialized data in the
top three bytes of the variable, effectively causing containsObject to
always return true.