[Network] Thread-safe TTNetworkRequestStarted/Stopped.
The TTNetworkRequestStarted/Stopped methods in the global network
header are now thread safe and throw an exception when you attempt
to stop more requests than you've started.

This does not solve the problem of not calling stop enough times.
A fix for this that doesn't involve a large amount of overengineering
is not immediately obvious.

Some considered alternatives for debugging:
 - A watchdog that periodically checks for active network connections.