Reduce include scope of TTDebugFlags.h
The fact that TTDebugFlags.h was included in TTGlobalCore.h meant that every
time you changed a debug flag, the entire project had to be rebuilt. This change
makes it so that only the files that use a debug flag have to be rebuilt, which
cuts build times and generally makes using these things more pleasant.