Allow collision keys in the XML parser in order to support RSS feeds.
RSS feeds tend to allow their feed items to be placed in the channel
object as flat items, instead of within an array. The original XML parser
implementation assumed that keys were always unique, but this is not the
case with RSS feeds. Options were added to make it possible to toggle this
functionality.