Deprecate the weak/strong methods in TTUserInfo.
Appears that Xcode might still be crashing in the wild with refactoring
due to these methods. To be safe, I'm deprecated the old methods and
providing new ones that don't conflict in names.

It should be noted that this fix still won't solve any existing crashes
that are still occurring. This is an intermediary change to allow people
to update to the new methods gracefully.