Add a new TTNavigator delegate method allowing modification of URLs to open
The new method is a superset of -navigator:shouldOpenURL: which allows the
navigator's delegate to prevent opening of a URL. The new method actually
returns the URL to open (or nil to prevent), which allows the delegate to modify
the URL to open.

I considered deprecating -navigator:shouldOpenURL: but I actually don't think
that's necessary since it's more convenient (and semantically clearer) to
implement that instead of -navigator:URLToOpen: if you don't need the
modification functionality.

Also, I have an Xcode plugin that strips trailing whitespace, which is why the
huge amount of changes to these files. To see the actual changes, do "git show
-b".