Make UIViewAdditions methods for showing as keyboard send right notifications
I recently changed TTViewController to correctly handle
UIKeyboard{Will,Did}{Show,Hide}Notifications, and without this change, table
view that autoresized for the keyboard were broken when some other view was
displayed as the keyboard. This makes the relevant code in UIViewAdditions send
the correct notifications so it really does act like a keyboard appearing and
disappearing.