Fix keyboard-disappearing resizing issue
This task of TTViewControllers auto-resizing for keyboard changes is
fraught with peril. In this instance, the problem was appearing in a
couple of places in the FB app. When logging in, the controller
displayed immediately after the login controller (the feed controller)
was getting -keyboardWillDisappear: (because it was disappearing from
the login screen) and was resizing itself to be the keyboard's height
taller than the screen. Another place was in search: when drilling
into a controller from a search listing with the keyboard up, the new
controller would again be the keyboard's height taller than the
screen. This change fixes that and also corrects a possible
double-initialization of the table view, resulting from accessor
properties actually being methods that initialize the ivar in
question. See the comment. >_<