Fix the popup view controller issue
Adds a method to UIViewController which allows it to express whether
it should ever be considered the "top" view controller, and added a
check to -[TTNavigator topViewController] to respect this. Defaults to
YES in UIViewController; the only override for now is
TTPopupViewController, which overrides to NO. In future, we can have
subclasses of TTPopupViewController re-override it to YES if need be.

Tested the camera button in the FB app's publisher bar (on the device)
and verified that the action sheet buttons do what they're supposed
to. Also tried the action sheet buttons for the Profile action sheet
(button in top right corner of friend's profile).

Review request: 73