[Core] Remove hard-coding in number format localization.
TTFormatInteger() was committing the cardinal sin of hardcoding a
localizable aspect of a number format. I traced it to a commit that
claimed this line of code was necessary to get the comma to show up,
which is NOT TRUE and actively works against localization of number
formatting. The grouping separator is not the same in all locales.

Testing: set system quantity formatting locale to en_US and
TTFormatInteger(1024) and make sure the result is 1,024. Set locale to
fr_FR and make sure the result is 1.024.