Remove JSON from the central Three20 source. Start to tear apart the libs.
This is the first commit of many that will begin to separate Three20
into multiple libraries. These changes won't have a direct impact
on the core Three20 library.

However, that being said, the core Three20 library will eventually
be phased out. We'll likely replace it with a library that simply
is dependent upon the four central libraries.