Remove all depencies on the legacy Three20 project. Create a clean version.
The new Three20 lib is fully modernized with the new modular system. The
old lib project still exists and is still backwards-compatible. Its
source is in Three20_Legacy.

The goal here is that people will switch to the new Three20 lib. The
new Three20 lib takes advantage of all of the great new modularization
and includes all of the original Three20 components (sans XML).