[Core] Use day name in short date strings only for dates <= 5 days ago
Summary:
Previously, when formatting a date, we would use the day name for any
date that was <= 60 * 24 * 7 seconds ago (one week's worth of
seconds).

This was really confusing for dates that were just short a week's worth
of seconds ago: if today is Friday, and the item we're formatting was
almost a week ago (but also on Friday), it would show up as "Fri".

Now, we'll use 5 days' worth of seconds as the cut-off to use
day names. That way, weekends will still work, but you won't see
"Fri" for things that happened last Friday.

Test Plan:
Using a table view with items with dates in it, confirmed
previously an item with a date of just short of 60 * 24 * 7 seconds
ago was rendered as "Fri XX:XX AM".

Applied my fix, now it's rendered as "May 7 XX:XX AM".