[Three20] Remove the composite lib from the Three20 project.
Compositing the library was causing build problems when making universal
binaries. It seems that, in general, compositing libraries together should
be avoided due to the strangeness of multiple architectures being linked
together, for one.

This is going to cause a bit of pain for anyone using the library due
to the fact that they will now need to add each library to their project
individually. If there is an alternate solution, it will be much preferred.