Change launcher's "highlight view" to cover entire screen
This makes it so that all the chrome will be darkened (and shielded from events)
like the launcher view.

Testing: highlight a launcher item and verify that everything (except the system
status bar) is covered by it. Tap on the circle and make sure the event goes
directly through. Tap somewhere other than the circle and make sure the event
does not go through and the highlight disappears.