Fix shadow-flipping bug in OS 3.2 and later
For some unknown reason, Apple changed the behavior of shadow drawing in
CoreGraphics to interpret the y coordinate of the shadow offset in the opposite
direction. I haven't found a programmatic way to detect this, so I had to go
with the monumentally skanky solution of a simple OS version check.
(cherry picked from commit 86f9c2df2e50233258359899dbf2d4e90396f503)