[Style] Real fix for Asian styled text layout issues
Summary:
Here's the correct fix for Asian styled text layout issues.

My previous workaround worked for a few cases where the text was mostly
Latin, but it wasn't the correct fix for the underlying problem.

The actual problem was that -[NSString drawInRect:withFont:] naively
defaults to UILineBreakModeWordWrap, even if it can't wrap in the
height it's given.

When using UILineBreakModeWordWrap, this method prefers to try to word
wrap (and fail) than to condense the text to fit on a single line,
which is how TTStyledLayout assumes -[NSString drawInRect:withFont:]
works.

To fix this issue, we now explicitly specify the behavior we want
(UILineBreakModeClip -- truncate if the text doesn't fit in the rect),
which convinces NSString to condense the text to fit inside the rect.

Testing:
Reproduced Korean truncation issue by visiting a profile
in simulator and on phone (OS 4).

Applied fix, confirmed in simulator and on phone that truncation issue
was fixed.

Reviewed By: oyamauchi