[UINavigator] Persist navigation stack when the app is backgrounded (multitasking)
The ApplicationWillTerminate notification may not be posted when the app is quit
while it's in the background, so we have to treat being backgrounded as if the
app is terminating.

Unfortunately, there is no way to behave gracefully on OS versions prior to 4.0
except by passing the notification name as a string (not as a constant which
points to said string). There's no way to stop the dynamic linker from trying to
resolve that symbol.