Removed hard-coded mappings of TTTableItems to TTTableViewCells to allow easier creation of new TTTableItem subclasses
TTTableViewDataSource now delegates to each TTTableItem to determine what type of TTTableViewCell to create or dequeue. This makes creation of new TTTableItem/TTTableViewCell subclasses easier and means that TTTableViewDataSource does not have to be subclassed to add new hard-coded mappings in tableView:cellClassForObject:.

Delegation to TTTableItem also opens up the possibility of dynamically creating different associated TTTableViewCells based on TTTableItem properties.

Note that non-TTTableItem-derived list items are still handled as a special case.