[UINavigator] Remove use of deprecated queryDictionaryUsingEncoding:.
The new method has been used in its place and the original functionality
of the navigator should still be in place. Duplicate URL keys will still
not be handled correctly because this change is merely meant to remove
the deprecation while preserving identical functionality. To correctly
handle multiple URL keys some thought will need to be invested in the
current design of the navigator.