[UI] Ensure TTSearchDisplayController hides the overlay view when searches complete.
This fixes a bug where the error overlay view is not hidden when the
search is cancelled.
This happens currently when you try to search without a network
connection. The error overlay remains after the search is dismissed.

I could fix this within the application by subclass viewWillDisappear on
the search controller; however, in order to synchronize the animations
with the search fade out and to handle this globally going forward I
think it's better to fix this in three20.

Test Plan:
verified that the overlay view is now hidden when the search
is cancelled