[UINavigator] Make the popover controller global.
This is an important change because popovers are globally active
objects. By making the methods static on the TTBaseNavigator class
we ensure an intuitive design pattern for accessing popovers.

In practice, any popover that is displayed, whether through Three20
or outside, should be assigned to the TTBaseNavigator so that it is
aware of the popover state.