[UI] Revert network request changes for the web view controller.
Adding support for network status isn't as simple as it might seem
for this controller. If the controller is popped off the stack,
for example, we'll never get enough request stopped messages to
stop the indicator, resulting in a permanent loading state for the
app.

Barring a proper solution to this problem, for now we'll leave it
to the controller to display the fact that it is loading.