TTScrollView, now can Center Page Index Animated.
New method setCenterPageIndex:animated: allows change the visible page with one optional animation. Also the centerPageAnimationDuration property allows the user to configure the duration of this animation.

Limitation (to be improved):
Only animate if the distance between the actual page and the informed one. Example: If is one page 1 and you inform page 3, will not animate.