[UI] don't swap datasources
This patch removes swapping away the datasource of a tableview while the model is not shown. Instead it just sets the tableView's hidden property to YES.

Removing and reading the datasource is pretty much just a bag of hurt. E.g. it can happen that during model changes, the datasource gets replaced. If this happens, the tableView's beginUpdates and endUpdates work with different datasources (and thus crashes pretty much all the time).