[BUG | FIXED] NSURLConnection delegate methods not called
When an asynchronous request loader is executed by the
_loaderQueueTimer once the connection starts the thread exits and the
NSURLConnection delegate methods are never be called, because as
described in the documentation:

"Messages to the delegate will be sent on the thread that calls this
method. By default, for the connection to work correctly, the calling
thread’s run loop must be operating in the default run loop mode."

This cause the request queue to "saturate" itself when it reaches the
kMaxConcurrentLoads request number. This cause all the network system
to stop working, because all the new request will queued and the loader
will be never notified that the old ones are ended.